Prevent scp-style arguments from being mangled
An argument like me@example.com:/tmp/ is not something we should convert
into a Windows path; Use the absence of a slash before the colon as a
tell-tale that it is *not* a POSIX path list (exception: if the part
left of the colon is "." or "..").

This fixes Git for Windows' t5602-clone-remote-exec.sh and
t5603-clone-dirname.sh.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>